// src/shared/api/generated/models/profileFieldValueResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileFieldValueResponseDtoValue } from "./profileFieldValueResponseDtoValue";

export interface ProfileFieldValueResponseDto {
  type: string;
  value: ProfileFieldValueResponseDtoValue;
}